refactor: move dto folder from api to application layer
